+++ kompot/anndata/__init__.py
@@ -7,6 +7,7 @@ from .differential_expression import de, compute_differential_expression
 from .impute import impute_expression
 from .utils import RunInfo, RunComparison, check_underrepresentation
 from .cleanup import cleanup, get_field_status
+from .fdr_utils import recompute_fdr, extract_null_distribution
 
 __all__ = [
     "de",
@@ -18,5 +19,7 @@ __all__ = [
     "RunComparison",
     "check_underrepresentation",
     "cleanup",
-    "get_field_status"
+    "get_field_status",
+    "recompute_fdr",
+    "extract_null_distribution",
 ]